Catch a wider range of number parsing exceptions
On my system (latest XCode), parse errors in std::stoi/std::stof trigger
a std::invalid_argument exception, which (incorrectly) is not caught by
the std::logic_error catch block. This is potentially a libc++
regression and causes many of our tests to fail. I've now changed the
code so that it is a bit more agnostic about what types of exceptions
are thrown by these methods.