Reload AD integrator Python package after variant change
test_ad_integrators.py runs the AD integrators with many different
variants. The main problem here is that these inherit from the C++ class
'Integrator', which is variant-specific! This means that changing the
variant of the renderer requires first flushing out those old type
declarations and loading in the right ones. (This was causing lots of
test failures)