The Enoki project was renamed to Dr.Jit
Lots of activity has taken place in the Enoki project during the last
two years. While the outer API has remained mostly the same, there is an
all-new system on the inside. Keeping the old name would have been quite
confusing -- the new name 'Dr.Jit' (i.e. a JIT for Differentiable
Rendering) signifies the scale of these changes. A doctor is somebody
that one can turn to when a situation is beyond one's control, which
seems somehow fitting given how hard implementing differentiable
rendering algorithms has become :-).

Note that this is a quite intrusive change which touches almost every
file of the renderer. A .git-blame-ignore-revs entry will be added in
a subsequent commit to avoid messing with the per-file history.