Better rgb/spectrum tag convertion in scene loading
This commit makes changes in the interpretation of <rgb value=".."> and <spectrum value=".."> properties
during scene loading for both RGB and spectral modes.

The main outcome of these changes:

1. In spectral mode, <spectrum ..> properties remain untouched (used to
be multiplied by D65 for emitters, and normalized according to CIE
curves)

2. In RGB mode, <spectrum ..> properties are pre-integrated and
normalized according to CIE curves. For reflectance properties this
pre-integration will take into account the D65 illuminant in order to
mimic the behavior of spectral mode.