astroobject: Minor implementation and test updates
This commit modifies the implementation of the `astroobject` plugin to
align it better with that of `envmap`. This should, hopefully, make the
implementation of missing methods and features easy, when time comes.

I also refactored the tests a bit so that they are a bit closer to an 
real actual use case.