Remove error on reads over MemoryStream's size.
This makes MemoryStream consistent with FileStream for reads beyond the maximum size. This is needed for the MemoryStream to be compatible with all image file formats. In particular, the JPEG library reads beyond the specified maximum buffer size. Previously, this meant that a JPEG could not easily be deserialized from a MemoryStream.