`RadianceMeter::sample_ray_differential` Make sure we JIT zero initialize `RayDifferential`
Proposed fix for #1196. When used as an inner sensor to the batch sensor, we still need `ray.o_x` and `ray.o_y` to be valid JIT variables even if they're not used (`ray.has_differentials = false`).

This is because here `sample_ray_differential` is now a vectorized call which checks that the return type of variables for each callable are valid, which in particular includes all variables within the traversed `RayDifferential` Dr.Jit struct