set_variant(): try next requested variant upon ImportError
If the user specifies several variants in order:
    mi.set_variant("cuda_ad_rgb", "llvm_ad_rgb")
and they are all compiled (available) but some fail to import,
then keep trying with the next requested variant instead of throwing an exception.

This could happen e.g. when requested a CUDA variant with CUDA installed but no GPU available,
or when LLVM is not installed.