Remove ``Thread``, ``ThreadEnvironment``, ScopedSetThreadEnvironment`` classes
Everything related to multithreading in Mitsuba is nowadays done by the
nanothread thread pool. This commit removes the ``Thread``,
``ThreadEnvironment``, ``ScopedSetThreadEnvironment`` classes that
were remainders of the Mitsuba 1/2 architecture.

Only a few methods remain for compatibility (but with deprecation
warnings). The class will be entirely removed in a future release.

- ``Thread.file_resolver()``, ``Thread.logger()``, ``Thread.thread()``,
  ``Thread.wait_for_tasks()``.

This commit also removes thread name printing from logger class.