Relax stub variant selection
The code that selects a variant for stub generation was somewhat strict
and could sometimes fail a build. This commit switches to a more
permissive system that scores variants to select the "best" variant. In
practice, this prioritizes variants with fancy spectral representations
(spectral rendering, polarization), where some types diverge that are
identical in simpler variants (e.g. `Spectrum` vs
`UnpolarizedSpectrum`).