Fetch sensor during `render` call for `stokes`
Previously, the sample() method always assumed that we used the scene's first sensor.

In addition, this fixes a subtle performance regression where accessing the sensor through `scene->sensors()[0]` could trigger a Python reference count change on the sensor object. This was an issue when using the scalar backend as their would be high contention over the GIL to chnage the reference count.