Faster importance sampling for the Sunsky model (#1800)
This PR refactors the way that the sky is importance sampled in the Sunsky emitters. Instead of using Truncated Gaussian Mixtures Models (TGMM), it now uses a composition of a downwards warp (towards the horizon) and a single truncated gaussian aligned on the sun's azimuth, yielding the following pdf:

<h3>$$f(u:=cos(\theta), v:=\phi) = w_{0} \cdot f_{1}(u; w_{1}) + (1- w_{0}) \cdot f_{1}(u, v; w_{2...4})$$</h3>
With:
<h3>$f_1(u; w_1) = \frac{1}{2 \pi} \cdot \frac{1}{w_1 \cdot (1+u^2(1/w_1^2-1))^{1.5}}$</h3>
And:
<h3>$f_2(u, v; w_{2...4})$</h3> 

Is a tructated gaussian defined on $[cos(\pi/2), cos(0)] \times [-\pi, \pi]$ and parametrized by: $\mu_{u} = w_{2}, \sigma_{u}=w_{3}$ and $\sigma_{v}=w_{4}$, with the mean for the azimuth is fixed at the sun's azimuth.

### Performance comparisons

**Values where the TGMM entry is the interpolation of 4 TGMMs** (current implementation)
|              | LLVM (#`gathers`)| CUDA (#`ld.global`)   |
| ---------- | ------------------------ | ---------------------------    |
| TGMM | 125                         | 124 + 1 texture read |
| MPDF  | 5                             | 7 texture reads          |

**Values where the TGMM entry is the result of clipping a single TGMM**
<img width="1570" height="495" alt="image" src="https://github.com/user-attachments/assets/073f09c0-f226-4a75-9336-b02a40432ab5" />
<img width="989" height="589" alt="image" src="https://github.com/user-attachments/assets/2ae825a1-f2cf-4bee-a6ab-9554d99d4e77" />
<img width="554" height="435" alt="image" src="https://github.com/user-attachments/assets/dfe0d90f-0855-4a1b-84ed-277923e752bc" />